imp difficulty finder
We may want to eventually compute aggregate statistics, like .mean() on the " Value" column. But right now, the values in that column are strings. Convert the" Value" column to floats. If you'd like to, you can create a new column with float values.

While most of the values in the " Value" column represent a dollar amount as a string, note that some do not — these values will need to be handled differently!

Now that you can filter the dataset of question, use your new column that contains the float values of each question to find the "difficulty" of certain topics. For example, what is the average value of questions that contain the word "King"?

Make sure to use the dataset that contains the float values as the dataset you use in your filtering function.