find unique answers
Write a function that returns the count of unique answers to all of the questions in a dataset. For example, after filtering the entire dataset to only questions containing the word "King", we could then find all of the unique answers to those questions. The answer "Henry VIII" appeared 55 times and was the most common answer.